Add algolia-ping function
